requirements.txt
python-dotenv
langchain
langchain-core
langchain-pinecone
pinecone
transformers
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
sentence-transformers